Organization for edit-item component
