+++ src/app/_components/edit-item/dialog-overview-example/dialog-overview-example.component.css
@@ -1,0 +1,23 @@
+
+mat-form-field {
+  font-family: "Lato", Arial, sans-serif;
+  color: red;
+  opacity: 0.6;
+  font-weight: 300;
+  font-size: 18px;
+}
+
+.mat-raised-button {
+  background-color: #c4ff8652;
+  color: rgba(33, 37, 41, 0.8);
+  font-family: "Lato", Arial, sans-serif;
+  font-weight: 400;
+  font-size: 16px;
+}
+
+.mat-input-element {
+  color: rgba(33, 37, 41, 0.8);
+  font-family: "Lato", Arial, sans-serif;
+  font-weight: 300;
+  font-size: 18px;
+}
