+++ src/app/_components/edit-item/index.ts
@@ -1,0 +1,3 @@
+export * from './edit-item.component';
+export * from './dialog-overview-example/dialog-overview-example.component';
+export * from './dialog-overview-example-dialog/dialog-overview-example-dialog.component';
