--- src/app/edit-item/edit-item.component.scss
@@ -1,38 +1,0 @@
-$light-grey: rgba(33, 37, 41, 0.18);
-$text-grey: rgba(33, 37, 41, 0.8);
-
-
-.grey-btn {
-  opacity: 0.5;
-  background-color: $light-grey;
-}
-
-.mat-raised-button {
-  background-color: #c4ff8652;
-  color: $text-grey;
-  font-family: "Lato", Arial, sans-serif;
-  font-weight: 400;
-  font-size: 16px;
-}
-
-.mat-input-element {
-  color: $text-grey;
-  font-family: "Lato", Arial, sans-serif;
-  font-weight: 300;
-  font-size: 18px;
-}
-
-//
-// <div class="mat-app-background">
-// <ol>
-// <li>
-//   <mat-form-field>
-//     <input matInput [(ngModel)]="name" placeholder="Enter a drug class">
-
-mat-form-field {
-  font-family: "Lato", Arial, sans-serif;
-  color: red;
-  opacity: 0.6;
-  font-weight: 300;
-  font-size: 18px;
-}
